Move editor map to separate source and header files
Move the `CEditorMap` and `CDataFileWriterFinishJob` classes to a separate header file and move the function definitions to source files.

Add an explicit constructor for the `CEditorMap` class to ensure that the editor pointer is always initialized.

Add a getter `Editor()` for the editor pointer of the `CEditorMap` class, to make the member variable private.

This is preparation for moving various functions and members from the `CEditor` to the `CEditorMap` class and eventually supporting multiple maps at the same time.
